# apps/backend/auth/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"AuthSignup": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"STUDENT",
							"TUTOR"
						]
					},
					"password": {
						"type": "string"
					},
					"googleId": {
						"type": "string"
					},
					"is_verified": {
						"type": "boolean"
					}
				},
				"required": [
					"firstname",
					"lastname",
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_AuthSignup_": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"STUDENT",
							"TUTOR"
						]
					},
					"password": {
						"type": "string"
					},
					"googleId": {
						"type": "string"
					},
					"is_verified": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"AuthSignupWithPicture": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_AuthSignup_"
					},
					{
						"properties": {
							"picture": {
								"type": "string",
								"nullable": true
							},
							"_id": {
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
			"Pick_AuthSignupWithPicture.firstname-or-lastname-or-picture-or-email-or-_id-or-role_": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"STUDENT",
							"TUTOR"
						]
					},
					"picture": {
						"type": "string"
					},
					"_id": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"UserPersional": {
				"$ref": "#/components/schemas/Pick_AuthSignupWithPicture.firstname-or-lastname-or-picture-or-email-or-_id-or-role_"
			},
			"Pick_AuthSignup.email-or-password_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"LoginTypes": {
				"$ref": "#/components/schemas/Pick_AuthSignup.email-or-password_"
			},
			"Partial_UserTypes_": {
				"properties": {
					"firstname": {
						"type": "string"
					},
					"lastname": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"USER",
							"STUDENT",
							"TUTOR"
						]
					},
					"password": {
						"type": "string"
					},
					"googleId": {
						"type": "string"
					},
					"is_verified": {
						"type": "boolean"
					},
					"_id": {
						"type": "string"
					},
					"picture": {
						"type": "string"
					},
					"authId": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"StudentType": {
				"allOf": [
					{
						"$ref": "#/components/schemas/Partial_UserTypes_"
					},
					{
						"properties": {
							"role": {
								"type": "string"
							},
							"studentCard": {
								"type": "string"
							},
							"grade": {
								"type": "number",
								"format": "double"
							},
							"education": {
								"type": "string"
							},
							"schoolName": {
								"type": "string"
							},
							"authId": {
								"type": "string"
							}
						},
						"required": [
							"studentCard",
							"grade",
							"education",
							"schoolName"
						],
						"type": "object"
					}
				]
			},
			"Student": {
				"properties": {
					"authId": {
						"type": "string"
					},
					"schoolName": {
						"type": "string"
					},
					"education": {
						"type": "string"
					},
					"grade": {
						"type": "number",
						"format": "double"
					},
					"studentCard": {
						"type": "string"
					}
				},
				"required": [
					"schoolName",
					"education",
					"grade",
					"studentCard"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "auth",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/auth-health": {
			"get": {
				"operationId": "CheckHealth",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthSignup"
							}
						}
					}
				}
			}
		},
		"/v1/auth/verify": {
			"get": {
				"operationId": "VerifySignupEmail",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"Data": {
											"$ref": "#/components/schemas/AuthSignupWithPicture"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"token",
										"Data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/password-reset/verify": {
			"get": {
				"operationId": "VerifyResetPasswordEmail",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/login": {
			"post": {
				"operationId": "LoginWithEmail",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"Data": {
											"$ref": "#/components/schemas/UserPersional"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"token",
										"Data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginTypes"
							}
						}
					}
				}
			}
		},
		"/v1/auth/google": {
			"get": {
				"operationId": "GoogleOAuth",
				"responses": {
					"200": {
						"description": "FOUND",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"redirectUrl": {}
									},
									"required": [
										"redirectUrl"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/google/callback": {
			"get": {
				"operationId": "GoogleOAuthData",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"Data": {
											"$ref": "#/components/schemas/AuthSignupWithPicture"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"token",
										"Data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/logout": {
			"get": {
				"operationId": "Logout",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"isLogout": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"isLogout",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/auth/password-reset/request": {
			"post": {
				"operationId": "RequestResetPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"email": {
										"type": "string"
									}
								},
								"required": [
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/reset-password": {
			"post": {
				"operationId": "ConfirmPassword",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "token",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									}
								},
								"required": [
									"password"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/v1/auth/identify": {
			"get": {
				"operationId": "IdentifyAuth",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"Data": {
											"$ref": "#/components/schemas/UserPersional"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"token",
										"Data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/v1/auth/v1/students/become-student": {
			"post": {
				"operationId": "BecomeToStudent",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"token": {
											"type": "string"
										},
										"Data": {
											"$ref": "#/components/schemas/StudentType"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"token",
										"Data",
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Student"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}